refactor(table-react): move types to separate file to fix circ.deps
affects: @fremtind/jkl-table-react
